test(epub): split the layout tests into their own file
layout.rs had grown to 1597 lines, well past the ~500 guideline, over half of it
tests. Move them to layout_tests.rs behind #[cfg(test)] #[path], the pattern
reader-core already uses for reflow_tests.rs: 883 lines of code, 696 of tests.

Collapse the repeated Block construction in those tests behind a runs() helper —
five hand-rolled Inline::Run(TextRun{..}) literals became one — so para() is now
styled_para() with a default style rather than a separate near-copy, and
styled_heading no longer destructures a paragraph to steal its content.

Also drop #[must_use] from effective_align, which is private.
