+++ inkread-epub/src/layout.rs
@@ -83,7 +83,6 @@ impl Align {
 /// declare for ordinary body prose (`p { text-align: justify }` is near-universal), and honouring
 /// those would silently override the setting the reader chose for the text they actually read. So
 /// they defer to `user`.
-#[must_use]
 fn effective_align(style: &BlockStyle, user: Align) -> Align {
     match style.align {
         Some(Align::Center) => Align::Center,
@@ -880,718 +879,5 @@ fn hyphenate_fit<'w>(
 }
 
 #[cfg(test)]
-mod tests {
-    #[test]
-    fn layout_digest_is_stable_and_sensitive_to_layout_fields() {
-        let base = LayoutOpts::new(400.0, 600.0, 16.0);
-        // Deterministic: identical opts → identical digest (same process AND across
-        // processes — the persisted cache key contract).
-        assert_eq!(
-            base.layout_digest(),
-            LayoutOpts::new(400.0, 600.0, 16.0).layout_digest()
-        );
-
-        // Every layout-affecting field flips the digest.
-        let d = base.layout_digest();
-        assert_ne!(
-            d,
-            LayoutOpts {
-                page_w: 401.0,
-                ..base
-            }
-            .layout_digest(),
-            "width"
-        );
-        assert_ne!(
-            d,
-            LayoutOpts {
-                page_h: 601.0,
-                ..base
-            }
-            .layout_digest(),
-            "height"
-        );
-        assert_ne!(
-            d,
-            LayoutOpts {
-                margin: base.margin + 1.0,
-                ..base
-            }
-            .layout_digest(),
-            "margin"
-        );
-        assert_ne!(
-            d,
-            LayoutOpts {
-                font_px: 17.0,
-                ..base
-            }
-            .layout_digest(),
-            "font"
-        );
-        assert_ne!(
-            d,
-            LayoutOpts {
-                line_spacing: 1.5,
-                ..base
-            }
-            .layout_digest(),
-            "line spacing"
-        );
-        assert_ne!(
-            d,
-            LayoutOpts {
-                para_gap: base.para_gap + 1.0,
-                ..base
-            }
-            .layout_digest(),
-            "para gap"
-        );
-        assert_ne!(
-            d,
-            LayoutOpts {
-                align: Align::Justify,
-                ..base
-            }
-            .layout_digest(),
-            "align"
-        );
-    }
-
-    #[test]
-    fn layout_digest_is_pinned_against_algorithm_drift() {
-        // The digest keys persisted paginations (ADR-INKREAD-0013 D1); pin a known value so a future
-        // change to the FNV constants/algorithm — which would silently orphan every cached
-        // pagination — is caught here, exactly as reader-core's identity fingerprint is pinned.
-        let opts = LayoutOpts {
-            page_w: 400.0,
-            page_h: 600.0,
-            margin: 24.0,
-            font_px: 16.0,
-            line_spacing: 1.4,
-            para_gap: 11.2,
-            align: Align::Left,
-        };
-        assert_eq!(opts.layout_digest(), 17_685_407_801_978_826_572);
-    }
-
-    use super::*;
-    use crate::content::{parse_blocks, TextRun};
-    use crate::css::BlockStyle;
-
-    /// Fixed-pitch metrics: every char advances `0.5 * size` (bold/italic ignored). Deterministic, so
-    /// wrapping/pagination can be asserted exactly without a font.
-    struct Mono;
-    impl Metrics for Mono {
-        fn advance(&self, text: &str, size_px: f32, _b: bool, _i: bool) -> f32 {
-            text.chars().count() as f32 * size_px * 0.5
-        }
-    }
-
-    fn para(text: &str) -> Block {
-        Block::Paragraph {
-            content: vec![Inline::Run(TextRun {
-                text: text.into(),
-                bold: false,
-                italic: false,
-                href: None,
-            })],
-            style: BlockStyle::default(),
-        }
-    }
-
-    /// Build the opts used by the #188 declared-style tests: 10px Mono font (5px/char), content
-    /// width 100 → 20 chars per line, and a reader whose global alignment is the default Left.
-    fn style_opts() -> LayoutOpts {
-        LayoutOpts {
-            page_w: 100.0,
-            page_h: 10_000.0,
-            margin: 0.0,
-            font_px: 10.0,
-            line_spacing: 1.0,
-            para_gap: 0.0,
-            align: Align::Left,
-        }
-    }
-
-    fn styled_para(text: &str, style: BlockStyle) -> Block {
-        let Block::Paragraph { content, .. } = para(text) else {
-            unreachable!()
-        };
-        Block::Paragraph { content, style }
-    }
-
-    fn styled_heading(text: &str, style: BlockStyle) -> Block {
-        let Block::Paragraph { content, .. } = para(text) else {
-            unreachable!()
-        };
-        Block::Heading {
-            level: 1,
-            content,
-            style,
-        }
-    }
-
-    /// #188: the title page rendered hard-left because `text-align: center` never reached layout.
-    #[test]
-    fn a_book_declared_centre_centres_even_when_the_reader_prefers_left() {
-        let opts = style_opts();
-        let style = BlockStyle {
-            align: Some(Align::Center),
-            ..Default::default()
-        };
-        let pages = paginate(&[styled_heading("PRIDE", style)], &opts, &Mono);
-        let x = pages[0].lines[0].runs[0].x;
-        // "PRIDE" at heading scale is narrower than the content box, so it must be inset.
-        assert!(x > 0.0, "declared centre was ignored: x = {x}");
-
-        // …and an undeclared heading still honours the reader's Left.
-        let plain = paginate(
-            &[styled_heading("PRIDE", BlockStyle::default())],
-            &opts,
-            &Mono,
-        );
-        assert_eq!(plain[0].lines[0].runs[0].x, 0.0);
-    }
-
-    /// The other half of the #188 policy: a book that justifies its prose must not override the
-    /// alignment the reader chose for the text they actually read.
-    #[test]
-    fn a_book_declared_justify_or_left_defers_to_the_reader() {
-        let opts = style_opts();
-        for declared in [Align::Justify, Align::Left] {
-            let style = BlockStyle {
-                align: Some(declared),
-                ..Default::default()
-            };
-            // A long paragraph so justification would visibly spread the first line's runs.
-            let text = "alpha bravo charlie delta echo foxtrot golf hotel";
-            let book = paginate(&[styled_para(text, style)], &opts, &Mono);
-            let user = paginate(&[styled_para(text, BlockStyle::default())], &opts, &Mono);
-            assert_eq!(
-                book[0].lines[0]
-                    .runs
-                    .iter()
-                    .map(|r| r.x)
-                    .collect::<Vec<_>>(),
-                user[0].lines[0]
-                    .runs
-                    .iter()
-                    .map(|r| r.x)
-                    .collect::<Vec<_>>(),
-                "book-declared {declared:?} should defer to the reader's Left"
-            );
-        }
-    }
-
-    /// The reader's own Justify still applies to blocks the book says nothing about — the policy
-    /// suppresses the *book's* left/justify, not the setting.
-    #[test]
-    fn the_readers_justify_still_applies_to_undeclared_blocks() {
-        let opts = LayoutOpts {
-            align: Align::Justify,
-            ..style_opts()
-        };
-        let text = "alpha bravo charlie delta echo foxtrot golf hotel";
-        let pages = paginate(&[styled_para(text, BlockStyle::default())], &opts, &Mono);
-        let first = &pages[0].lines[0];
-        assert!(first.runs.len() > 1);
-        let last_x = first.runs.last().unwrap().x;
-        let flush = paginate(
-            &[styled_para(text, BlockStyle::default())],
-            &LayoutOpts {
-                align: Align::Left,
-                ..style_opts()
-            },
-            &Mono,
-        );
-        assert!(
-            last_x > flush[0].lines[0].runs.last().unwrap().x,
-            "justification did not spread the line"
-        );
-    }
-
-    /// #188 / #163: the paragraph indent is applied even to blocks whose stylesheet turns it off.
-    #[test]
-    fn a_declared_zero_text_indent_drops_the_first_line_indent() {
-        let opts = style_opts();
-        let indented = paginate(
-            &[styled_para("short line", BlockStyle::default())],
-            &opts,
-            &Mono,
-        );
-        assert_eq!(indented[0].lines[0].runs[0].x, 12.0, "1.2em of 10px");
-
-        let style = BlockStyle {
-            indent: Some(false),
-            ..Default::default()
-        };
-        let flat = paginate(&[styled_para("short line", style)], &opts, &Mono);
-        assert_eq!(
-            flat[0].lines[0].runs[0].x, 0.0,
-            "text-indent: 0 was ignored"
-        );
-    }
-
-    /// A centred block must not also carry the synthetic prose indent, which would push it off
-    /// centre by half the indent.
-    #[test]
-    fn a_centred_paragraph_drops_the_indent_it_would_otherwise_carry() {
-        let opts = style_opts();
-        let centred = BlockStyle {
-            align: Some(Align::Center),
-            ..Default::default()
-        };
-        let text = "abcd"; // 4 chars * 5px = 20px wide in a 100px box
-        let pages = paginate(&[styled_para(text, centred)], &opts, &Mono);
-        assert_eq!(
-            pages[0].lines[0].runs[0].x, 40.0,
-            "expected (100 - 20) / 2 with no indent skewing it"
-        );
-    }
-
-    /// #188: headings were laid out with `bold_all = true` unconditionally.
-    #[test]
-    fn a_declared_normal_font_weight_unbolds_a_heading() {
-        let opts = style_opts();
-        let bold = paginate(
-            &[styled_heading("Title", BlockStyle::default())],
-            &opts,
-            &Mono,
-        );
-        assert!(bold[0].lines[0].runs[0].bold, "headings default to bold");
-
-        let style = BlockStyle {
-            bold: Some(false),
-            ..Default::default()
-        };
-        let normal = paginate(&[styled_heading("Title", style)], &opts, &Mono);
-        assert!(
-            !normal[0].lines[0].runs[0].bold,
-            "font-weight: normal was ignored"
-        );
-
-        // …and the converse: a paragraph the book bolds comes out bold.
-        let style = BlockStyle {
-            bold: Some(true),
-            ..Default::default()
-        };
-        let p = paginate(&[styled_para("x", style)], &opts, &Mono);
-        assert!(p[0].lines[0].runs[0].bold);
-    }
-
-    #[test]
-    fn long_paragraph_wraps_to_multiple_lines() {
-        // 10px font → 5px/char. content_w = 100 → 20 chars/line. 60-char paragraph → 3 lines.
-        let opts = LayoutOpts {
-            page_w: 100.0 + 2.0 * 0.0,
-            page_h: 10_000.0,
-            margin: 0.0,
-            font_px: 10.0,
-            line_spacing: 1.0,
-            para_gap: 0.0,
-            align: Align::Left,
-        };
-        let words = "aaaa ".repeat(12); // 12 words of 4 chars → ~ wraps
-        let pages = paginate(&[para(words.trim())], &opts, &Mono);
-        assert_eq!(pages.len(), 1);
-        assert!(
-            pages[0].lines.len() >= 3,
-            "wrapped: {}",
-            pages[0].lines.len()
-        );
-        // No run exceeds the content width.
-        for line in &pages[0].lines {
-            for r in &line.runs {
-                let w = r.x + r.text.chars().count() as f32 * 5.0;
-                assert!(w <= 100.0 + 0.01, "run overflows: {w}");
-            }
-        }
-    }
-
-    /// The narrow test viewport for the unspaced-script (CJK) wrapping tests: 10px font → 5px/char
-    /// (Mono); content 50px. The paragraph's 12px (1.2em) indent applies to the **first line only**,
-    /// so 7 chars fit there and 10 on every line after it.
-    fn cjk_opts() -> LayoutOpts {
-        LayoutOpts {
-            page_w: 50.0,
-            page_h: 10_000.0,
-            margin: 0.0,
-            font_px: 10.0,
-            line_spacing: 1.0,
-            para_gap: 0.0,
-            align: Align::Left,
-        }
-    }
-
-    /// Each line's runs joined (the CJK tests place one run per line).
-    fn line_texts(pages: &[Page]) -> Vec<String> {
-        pages
-            .iter()
-            .flat_map(|p| &p.lines)
-            .map(|l| l.runs.iter().map(|r| r.text.as_str()).collect())
-            .collect()
-    }
-
-    /// #163: the first-line indent must land on the first line and nowhere else.
-    ///
-    /// This typography omits the blank line between paragraphs on purpose, so the indent is the
-    /// *only* thing marking where one paragraph ends and the next begins. Applying it to every line
-    /// — which is what the code did — left prose with no paragraph breaks at all, and spent 1.2em of
-    /// every line's width on nothing.
-    #[test]
-    fn only_a_paragraphs_first_line_is_indented() {
-        let pages = paginate(
-            &[para("alpha bravo charlie delta echo foxtrot")],
-            &cjk_opts(),
-            &Mono,
-        );
-        let lines: Vec<&LayoutLine> = pages.iter().flat_map(|p| &p.lines).collect();
-        assert!(
-            lines.len() > 1,
-            "need a wrapped paragraph to test: {lines:?}"
-        );
-
-        assert_eq!(
-            lines[0].runs[0].x, 12.0,
-            "the first line carries the 1.2em indent"
-        );
-        for (i, line) in lines.iter().enumerate().skip(1) {
-            assert_eq!(line.runs[0].x, 0.0, "line {i} must start flush left");
-        }
-    }
-
-    /// The counterpart: a list item's indent is a *hanging* indent, so it applies to every line and
-    /// the marker hangs to the left of the first. The two must not be conflated.
-    #[test]
-    fn a_list_items_body_stays_indented_on_every_line() {
-        let item = Block::ListItem {
-            ordered: false,
-            index: 1,
-            content: vec![Inline::Run(TextRun {
-                text: "alpha bravo charlie delta echo".into(),
-                bold: false,
-                italic: false,
-                href: None,
-            })],
-            style: BlockStyle::default(),
-        };
-        let pages = paginate(&[item], &cjk_opts(), &Mono);
-        let lines: Vec<&LayoutLine> = pages.iter().flat_map(|p| &p.lines).collect();
-        assert!(lines.len() > 1, "need a wrapped item to test");
-
-        // The marker hangs at the content origin; the body starts indented past it.
-        assert_eq!(lines[0].runs[0].text, "•");
-        assert_eq!(lines[0].runs[0].x, 0.0, "the marker hangs left");
-        let body_x = lines[0].runs[1].x;
-        assert!(body_x > 0.0, "the body is inset past the marker");
-        for (i, line) in lines.iter().enumerate().skip(1) {
-            assert_eq!(
-                line.runs[0].x, body_x,
-                "wrapped line {i} stays under the body, not the marker"
-            );
-        }
-    }
-
-    #[test]
-    fn cjk_paragraph_wraps_between_characters() {
-        // No spaces, so the whole paragraph is one token — before UAX #14 it overflowed as a single
-        // unbreakable "word". 22 Han chars wrap greedily to 7 (the indented first line) + 10 + 5,
-        // nothing lost.
-        let text = "书山有路勤为径学海无涯苦作舟读万卷书行万里路";
-        let pages = paginate(&[para(text)], &cjk_opts(), &Mono);
-        let lines = line_texts(&pages);
-        assert_eq!(lines.len(), 3, "22 chars as 7+10+5: {lines:?}");
-        assert_eq!(
-            lines[0].chars().count(),
-            7,
-            "the indented first line is narrower"
-        );
-        assert_eq!(
-            lines[1].chars().count(),
-            10,
-            "later lines get the full column"
-        );
-        assert_eq!(lines.concat(), text, "no characters dropped or reordered");
-        // No line overflows the content box, indented or not.
-        for line in pages.iter().flat_map(|p| &p.lines) {
-            for r in &line.runs {
-                assert!(r.x + r.text.chars().count() as f32 * 5.0 <= 50.0 + 0.01);
-            }
-        }
-    }
-
-    #[test]
-    fn cjk_break_honors_kinsoku_and_anchors_stay_font_invariant() {
-        // Naive 7-chars-per-line breaking would start line 2 with 。— UAX #14 forbids a break
-        // before a closing form, so the break retreats to after 六 and 。hangs onto 七's line.
-        let pages = paginate(&[para("一二三四五六七。八九十")], &cjk_opts(), &Mono);
-        let lines = line_texts(&pages);
-        assert_eq!(lines, ["一二三四五六", "七。八九十"]);
-        assert!(lines.iter().all(|l| !l.starts_with('。')));
-        // The continuation run keeps a chapter-relative anchor = chars consumed before it
-        // (ADR-INKREAD-0012 — a highlight re-resolves across reflow).
-        let runs: Vec<&PlacedRun> = pages
-            .iter()
-            .flat_map(|p| &p.lines)
-            .flat_map(|l| &l.runs)
-            .collect();
-        assert_eq!(runs[0].anchor.char_offset, 0);
-        assert_eq!(runs[1].anchor.char_offset, 6);
-    }
-
-    #[test]
-    fn mixed_latin_cjk_breaks_at_the_script_boundary_opportunities() {
-        // "Hello你好世界" is one token (no spaces). Latin-internal positions are not eligible —
-        // only UAX #14 opportunities adjacent to an unspaced-script char — so the line fills to
-        // "Hello你好" (7 chars, 35px ≤ 38px) and wraps cleanly, no overflow, no hyphen.
-        let pages = paginate(&[para("Hello你好世界")], &cjk_opts(), &Mono);
-        let lines = line_texts(&pages);
-        assert_eq!(lines, ["Hello你好", "世界"]);
-        assert!(!lines[0].ends_with('-'), "no hyphen for an unspaced break");
-    }
-
-    #[test]
-    fn content_overflow_breaks_into_pages() {
-        // line height 10px, content_h 30px → 3 lines/page. 7 short paragraphs → 3 pages.
-        let opts = LayoutOpts {
-            page_w: 1000.0,
-            page_h: 30.0,
-            margin: 0.0,
-            font_px: 10.0,
-            line_spacing: 1.0,
-            para_gap: 0.0,
-            align: Align::Left,
-        };
-        let blocks: Vec<Block> = (0..7).map(|_| para("x")).collect();
-        let pages = paginate(&blocks, &opts, &Mono);
-        assert_eq!(pages.len(), 3, "7 lines / 3 per page");
-        assert_eq!(pages[0].lines.len(), 3);
-        assert_eq!(pages[2].lines.len(), 1);
-    }
-
-    #[test]
-    fn heading_uses_a_larger_line_height() {
-        let opts = LayoutOpts {
-            page_w: 10_000.0,
-            page_h: 10_000.0,
-            margin: 0.0,
-            font_px: 10.0,
-            line_spacing: 1.0,
-            para_gap: 0.0,
-            align: Align::Left,
-        };
-        let pages = paginate(
-            &[Block::Heading {
-                level: 1,
-                content: vec![Inline::Run(TextRun {
-                    text: "Title".into(),
-                    bold: false,
-                    italic: false,
-                    href: None,
-                })],
-                style: BlockStyle::default(),
-            }],
-            &opts,
-            &Mono,
-        );
-        let line = &pages[0].lines[0];
-        assert_eq!(line.height, 18.0, "h1 = 1.8 * 10"); // heading_scale(1)=1.8
-        assert!(line.runs[0].bold, "headings render bold");
-    }
-
-    #[test]
-    fn list_item_has_marker_and_hanging_indent() {
-        let opts = LayoutOpts::new(1000.0, 1000.0, 10.0);
-        let pages = paginate(
-            &[Block::ListItem {
-                ordered: true,
-                index: 3,
-                content: vec![Inline::Run(TextRun {
-                    text: "item text".into(),
-                    bold: false,
-                    italic: false,
-                    href: None,
-                })],
-                style: BlockStyle::default(),
-            }],
-            &opts,
-            &Mono,
-        );
-        let runs = &pages[0].lines[0].runs;
-        assert_eq!(runs[0].text, "3.", "ordered marker");
-        assert_eq!(runs[0].x, 0.0, "marker at content origin");
-        assert!(runs[1].x > 0.0, "body hangs past the marker");
-    }
-
-    #[test]
-    fn integrates_with_phase2_parsing() {
-        let blocks = parse_blocks("<html><body><h2>Hi</h2><p>one two three</p></body></html>");
-        let opts = LayoutOpts::new(400.0, 600.0, 16.0);
-        let pages = paginate(&blocks, &opts, &Mono);
-        assert!(!pages.is_empty());
-        let all_text: String = pages[0]
-            .lines
-            .iter()
-            .flat_map(|l| l.runs.iter())
-            .map(|r| r.text.clone())
-            .collect::<Vec<_>>()
-            .join(" ");
-        assert!(all_text.contains("Hi") && all_text.contains("three"));
-    }
-
-    #[test]
-    fn empty_blocks_make_no_pages() {
-        assert!(paginate(&[], &LayoutOpts::new(400.0, 600.0, 16.0), &Mono).is_empty());
-    }
-
-    /// Collect every placed run as `(block, char_offset, text)` in reading order.
-    fn run_anchors(pages: &[Page]) -> Vec<(usize, usize, String)> {
-        pages
-            .iter()
-            .flat_map(|p| p.lines.iter())
-            .flat_map(|l| l.runs.iter())
-            .map(|r| (r.anchor.block, r.anchor.char_offset, r.text.clone()))
-            .collect()
-    }
-
-    fn wide(font_px: f32) -> LayoutOpts {
-        LayoutOpts {
-            page_w: 100_000.0,
-            page_h: 100_000.0,
-            margin: 0.0,
-            font_px,
-            line_spacing: 1.0,
-            para_gap: 0.0,
-            align: Align::Left,
-        }
-    }
-
-    #[test]
-    fn run_anchors_track_chapter_character_offsets() {
-        // "alpha"(5)@0, space@5, "beta"(4)@6, space@10, "gamma"@11.
-        let pages = paginate(&[para("alpha beta gamma")], &wide(10.0), &Mono);
-        assert_eq!(
-            run_anchors(&pages),
-            vec![
-                (0, 0, "alpha".into()),
-                (0, 6, "beta".into()),
-                (0, 11, "gamma".into()),
-            ]
-        );
-    }
-
-    #[test]
-    fn block_index_increments_and_offset_continues_across_blocks() {
-        // block 0 "one two": one@0, two@4 → cursor 7; block 1 "three"@7.
-        let pages = paginate(&[para("one two"), para("three")], &wide(10.0), &Mono);
-        assert_eq!(
-            run_anchors(&pages),
-            vec![
-                (0, 0, "one".into()),
-                (0, 4, "two".into()),
-                (1, 7, "three".into()),
-            ]
-        );
-    }
-
-    #[test]
-    fn list_marker_shares_body_offset_and_does_not_consume_budget() {
-        let pages = paginate(
-            &[
-                Block::ListItem {
-                    ordered: true,
-                    index: 1,
-                    content: vec![Inline::Run(TextRun {
-                        text: "first".into(),
-                        bold: false,
-                        italic: false,
-                        href: None,
-                    })],
-                    style: BlockStyle::default(),
-                },
-                para("after"),
-            ],
-            &LayoutOpts::new(1000.0, 1000.0, 10.0),
-            &Mono,
-        );
-        let anchors = run_anchors(&pages);
-        // Marker "1." and body "first" both anchor at offset 0 of block 0; the marker adds no budget,
-        // so "after" (block 1) starts at 5 (= len("first")), not 7.
-        assert_eq!(anchors[0], (0, 0, "1.".into()), "marker shares body start");
-        assert_eq!(anchors[1], (0, 0, "first".into()), "body at block start");
-        assert_eq!(
-            anchors[2],
-            (1, 5, "after".into()),
-            "marker consumed no offset"
-        );
-    }
-
-    /// A test hyphenator allowing breaks at fixed byte offsets (regardless of the word).
-    struct HyphenAt(Vec<usize>);
-    impl Hyphenator for HyphenAt {
-        fn opportunities(&self, _word: &str) -> Vec<usize> {
-            self.0.clone()
-        }
-    }
-
-    fn narrow_para() -> LayoutOpts {
-        // Mono 10px → 5px/char; content 60 wide, paragraph indent 12 → 48px usable.
-        LayoutOpts {
-            page_w: 60.0,
-            page_h: 10_000.0,
-            margin: 0.0,
-            font_px: 10.0,
-            line_spacing: 1.0,
-            para_gap: 0.0,
-            align: Align::Left,
-        }
-    }
-
-    #[test]
-    fn long_word_hyphenates_and_suffix_keeps_its_anchor() {
-        // "hyphenation" (11 chars = 55px) overflows the 48px column; a break after 5 bytes fits
-        // ("hyphe" 25px + hyphen 5px = 30 ≤ 48), so it splits and the suffix anchors after the prefix.
-        let pages = paginate_with(
-            &[para("hyphenation")],
-            &narrow_para(),
-            &Mono,
-            &HyphenAt(vec![5]),
-        );
-        let runs: Vec<_> = pages[0].lines.iter().flat_map(|l| l.runs.iter()).collect();
-        assert_eq!(runs[0].text, "hyphe-", "prefix carries a trailing hyphen");
-        assert_eq!(runs[1].text, "nation", "suffix continues on the next line");
-        assert_eq!(runs[0].anchor.char_offset, 0);
-        assert_eq!(
-            runs[1].anchor.char_offset, 5,
-            "suffix anchored at prefix length (reflow-stable)"
-        );
-    }
-
-    #[test]
-    fn default_paginate_never_hyphenates() {
-        // The default ([`NoHyphen`]) places an overflowing word whole, on its own line.
-        let pages = paginate(&[para("hyphenation")], &narrow_para(), &Mono);
-        let runs: Vec<_> = pages[0].lines.iter().flat_map(|l| l.runs.iter()).collect();
-        assert_eq!(runs.len(), 1);
-        assert_eq!(runs[0].text, "hyphenation");
-    }
-
-    #[test]
-    fn run_anchors_are_font_size_invariant() {
-        // Wrapping differs by size, but each word keeps its (block, char_offset) — the reflow-stable
-        // property a highlight/Digest anchor relies on (ADR-INKREAD-0012).
-        let blocks = [
-            para("the quick brown fox jumps over"),
-            para("the lazy dog sleeps soundly"),
-        ];
-        let narrow = |fp: f32| {
-            let opts = LayoutOpts {
-                page_w: 60.0,
-                ..wide(fp)
-            };
-            run_anchors(&paginate(&blocks, &opts, &Mono))
-        };
-        assert_eq!(narrow(10.0), narrow(20.0));
-    }
-}
+#[path = "layout_tests.rs"]
+mod tests;
